# openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/auth/code": {
      "post": {
        "operationId": "AuthController_sendCode",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "phone": {
                    "type": "string",
                    "minLength": 8,
                    "maxLength": 24
                  },
                  "channel": {
                    "default": "telegram",
                    "type": "string",
                    "enum": [
                      "telegram",
                      "whatsapp",
                      "sms"
                    ]
                  },
                  "locale": {
                    "type": "string",
                    "enum": [
                      "ru",
                      "hy",
                      "en"
                    ]
                  }
                },
                "required": [
                  "phone"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "challengeId": {
                      "type": "string"
                    },
                    "resendAfter": {
                      "type": "number"
                    },
                    "expiresIn": {
                      "type": "number"
                    },
                    "channel": {
                      "type": "string",
                      "enum": [
                        "telegram",
                        "whatsapp",
                        "sms"
                      ]
                    }
                  },
                  "required": [
                    "challengeId",
                    "resendAfter",
                    "expiresIn",
                    "channel"
                  ]
                }
              }
            }
          }
        },
        "summary": "Отправить код входа (4 цифры, 5 мин, повтор через 60 с). Ответ одинаков для любого номера.",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/verify": {
      "post": {
        "operationId": "AuthController_verify",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "phone": {
                    "type": "string",
                    "minLength": 8,
                    "maxLength": 24
                  },
                  "code": {
                    "type": "string",
                    "pattern": "^\\d{4}$"
                  },
                  "app": {
                    "default": "client",
                    "type": "string",
                    "enum": [
                      "client",
                      "business"
                    ]
                  },
                  "name": {
                    "type": "string",
                    "maxLength": 120
                  },
                  "consent": {
                    "type": "boolean"
                  },
                  "locale": {
                    "type": "string",
                    "enum": [
                      "ru",
                      "hy",
                      "en"
                    ]
                  }
                },
                "required": [
                  "phone",
                  "code"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        },
                        "phone": {
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "locale": {
                          "type": "string",
                          "enum": [
                            "ru",
                            "hy",
                            "en"
                          ]
                        }
                      },
                      "required": [
                        "id",
                        "name",
                        "phone",
                        "locale"
                      ]
                    },
                    "app": {
                      "type": "string",
                      "enum": [
                        "client",
                        "business",
                        "platform"
                      ]
                    },
                    "mode": {
                      "type": "string",
                      "enum": [
                        "client",
                        "business"
                      ]
                    },
                    "activeBusinessId": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "memberships": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "businessId": {
                            "type": "string"
                          },
                          "businessName": {
                            "type": "string"
                          },
                          "kind": {
                            "type": "string"
                          },
                          "networkId": {
                            "type": [
                              "string",
                              "null"
                            ]
                          },
                          "staffId": {
                            "type": "string"
                          },
                          "role": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "businessId",
                          "businessName",
                          "kind",
                          "networkId",
                          "staffId",
                          "role"
                        ]
                      }
                    },
                    "staffLogin": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "mustChangePassword": {
                      "type": "boolean"
                    },
                    "consent": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "user",
                    "app",
                    "mode",
                    "activeBusinessId",
                    "memberships",
                    "staffLogin",
                    "mustChangePassword",
                    "consent"
                  ]
                }
              }
            }
          }
        },
        "summary": "Проверить код и войти (клиент или кабинет бизнеса). Ставит httpOnly cookie сессии.",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/password": {
      "post": {
        "operationId": "AuthController_password",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "login": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 64
                  },
                  "password": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 128
                  },
                  "channel": {
                    "type": "string",
                    "enum": [
                      "telegram",
                      "whatsapp",
                      "sms"
                    ]
                  }
                },
                "required": [
                  "login",
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Вход администратора логином и паролем (F-00-034). Если включён второй шаг — { secondFactor }.",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/second-factor": {
      "post": {
        "operationId": "AuthController_secondFactor",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "challengeId": {
                    "type": "string",
                    "minLength": 4,
                    "maxLength": 32
                  },
                  "code": {
                    "type": "string",
                    "pattern": "^\\d{4}$"
                  }
                },
                "required": [
                  "challengeId",
                  "code"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        },
                        "phone": {
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "locale": {
                          "type": "string",
                          "enum": [
                            "ru",
                            "hy",
                            "en"
                          ]
                        }
                      },
                      "required": [
                        "id",
                        "name",
                        "phone",
                        "locale"
                      ]
                    },
                    "app": {
                      "type": "string",
                      "enum": [
                        "client",
                        "business",
                        "platform"
                      ]
                    },
                    "mode": {
                      "type": "string",
                      "enum": [
                        "client",
                        "business"
                      ]
                    },
                    "activeBusinessId": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "memberships": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "businessId": {
                            "type": "string"
                          },
                          "businessName": {
                            "type": "string"
                          },
                          "kind": {
                            "type": "string"
                          },
                          "networkId": {
                            "type": [
                              "string",
                              "null"
                            ]
                          },
                          "staffId": {
                            "type": "string"
                          },
                          "role": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "businessId",
                          "businessName",
                          "kind",
                          "networkId",
                          "staffId",
                          "role"
                        ]
                      }
                    },
                    "staffLogin": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "mustChangePassword": {
                      "type": "boolean"
                    },
                    "consent": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "user",
                    "app",
                    "mode",
                    "activeBusinessId",
                    "memberships",
                    "staffLogin",
                    "mustChangePassword",
                    "consent"
                  ]
                }
              }
            }
          }
        },
        "summary": "Второй шаг входа по паролю: код на телефон (F-15-159)",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/password/change": {
      "post": {
        "operationId": "AuthController_changePassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "oldPassword": {
                    "type": "string",
                    "maxLength": 128
                  },
                  "newPassword": {
                    "type": "string",
                    "maxLength": 128
                  }
                },
                "required": [
                  "newPassword"
                ]
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Сменить пароль администратора; при первом входе старый не нужен (F-00-034)",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/mode": {
      "put": {
        "operationId": "AuthController_mode",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "mode": {
                    "type": "string",
                    "enum": [
                      "client",
                      "business"
                    ]
                  },
                  "businessId": {
                    "type": "string",
                    "maxLength": 32
                  }
                },
                "required": [
                  "mode"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        },
                        "phone": {
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "locale": {
                          "type": "string",
                          "enum": [
                            "ru",
                            "hy",
                            "en"
                          ]
                        }
                      },
                      "required": [
                        "id",
                        "name",
                        "phone",
                        "locale"
                      ]
                    },
                    "app": {
                      "type": "string",
                      "enum": [
                        "client",
                        "business",
                        "platform"
                      ]
                    },
                    "mode": {
                      "type": "string",
                      "enum": [
                        "client",
                        "business"
                      ]
                    },
                    "activeBusinessId": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "memberships": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "businessId": {
                            "type": "string"
                          },
                          "businessName": {
                            "type": "string"
                          },
                          "kind": {
                            "type": "string"
                          },
                          "networkId": {
                            "type": [
                              "string",
                              "null"
                            ]
                          },
                          "staffId": {
                            "type": "string"
                          },
                          "role": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "businessId",
                          "businessName",
                          "kind",
                          "networkId",
                          "staffId",
                          "role"
                        ]
                      }
                    },
                    "staffLogin": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "mustChangePassword": {
                      "type": "boolean"
                    },
                    "consent": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "user",
                    "app",
                    "mode",
                    "activeBusinessId",
                    "memberships",
                    "staffLogin",
                    "mustChangePassword",
                    "consent"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Переключатель «Я клиент / Мой бизнес» (В-21)",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/v1/auth/logout-all": {
      "post": {
        "operationId": "AuthController_logoutAll",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "keepCurrent": {
                    "default": true,
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "«Завершить все сеансы» (F-15-152, F-10-125); keepCurrent — кроме этого устройства",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/platform/login": {
      "post": {
        "operationId": "AuthController_platformLogin",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "login": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 64
                  },
                  "password": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 128
                  },
                  "channel": {
                    "type": "string",
                    "enum": [
                      "telegram",
                      "whatsapp",
                      "sms"
                    ]
                  }
                },
                "required": [
                  "login",
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "secondFactor": {
                      "type": "object",
                      "properties": {
                        "challengeId": {
                          "type": "string"
                        },
                        "phoneMasked": {
                          "type": "string"
                        },
                        "resendAfter": {
                          "type": "number"
                        },
                        "expiresIn": {
                          "type": "number"
                        }
                      },
                      "required": [
                        "challengeId",
                        "phoneMasked",
                        "resendAfter",
                        "expiresIn"
                      ]
                    }
                  },
                  "required": [
                    "secondFactor"
                  ]
                }
              }
            }
          }
        },
        "summary": "Панель: логин + пароль → код на телефон (всегда второй шаг)",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/platform/verify": {
      "post": {
        "operationId": "AuthController_platformVerify",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "challengeId": {
                    "type": "string",
                    "minLength": 4,
                    "maxLength": 32
                  },
                  "code": {
                    "type": "string",
                    "pattern": "^\\d{4}$"
                  }
                },
                "required": [
                  "challengeId",
                  "code"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        },
                        "phone": {
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "locale": {
                          "type": "string",
                          "enum": [
                            "ru",
                            "hy",
                            "en"
                          ]
                        }
                      },
                      "required": [
                        "id",
                        "name",
                        "phone",
                        "locale"
                      ]
                    },
                    "login": {
                      "type": "string"
                    },
                    "role": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "user",
                    "login",
                    "role"
                  ]
                }
              }
            }
          }
        },
        "summary": "Панель: код второго шага → сессия платформы",
        "tags": [
          "auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "BookTime API",
    "description": "Сервер для booking-platform. Ошибки — { code, message } (docs/backend/02-api.md §0).",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
...
}